requirements.txt
flask
python-novaclient
python-neutronclient
python-glanceclient
keystoneauth1
python-heatclient
networking-sfc
